Migrations bugfix

+++ migrations/m200109_141741_news2amp.php
@@ -12,10 +12,8 @@ class m200109_141741_news2amp extends Migration
      */
     public function safeUp()
     {
-
         if (is_null($this->getDb()->getSchema()->getTableSchema('{{%news}}')->getColumn('in_amp')))
             $this->addColumn('{{%news}}', 'in_amp', $this->boolean()->defaultValue(true)->after('source'));
-
     }
 
     /**
@@ -23,6 +21,7 @@ class m200109_141741_news2amp extends Migration
      */
     public function safeDown()
     {
-        $this->dropColumn('{{%news}}', 'in_amp');
+        if ($this->getDb()->getSchema()->getTableSchema('{{%news}}')->getColumn('in_amp'))
+            $this->dropColumn('{{%news}}', 'in_amp');
     }
 }
